Create Blackjack Game (1)
Currently a work in progress. I want to condense my hit_or_stand function, there's too many conditionals.  I also want to optimize this program as certain conditions output the undesired result. For instance, if the dealer or player has a hand that is valued more than 21,  they should both bust and end in a tie. Instead, it just gives the win to the opponent that has the higher deck value.